API review feedback (#988)

+++ sdk/inc/azure/core/internal/az_retry_internal.h
@@ -9,7 +9,7 @@
 #include <azure/core/_az_cfg_prefix.h>
 
 AZ_NODISCARD AZ_INLINE int32_t _az_retry_calc_delay(
-    int16_t attempt,
+    int32_t attempt,
     int32_t retry_delay_msec,
     int32_t max_retry_delay_msec)
 {
